+++ src/produto/produto-controller.ts
@@ -1,11 +1,19 @@
 import * as ProdutoService from "./produto-service";
 import * as RestauranteService from "../restaurante/restaurante-service";
+import * as jwt from "../auth/jwt-auth";
 import { Request, Response } from "express";
 
 export class ProdutoController {
 
     async index(req: Request, res: Response): Promise<Response> {
         try {
+
+            if(!req.headers.authorization.split(' ')[1]) return res.status(403).json({ error: "Sem autorização!" });
+
+            const validate = jwt.verify(req.headers.authorization.split(' ')[1]);
+
+            if(await validate == null) return res.status(403).json({ error: "Sem autorização!" });
+
             const produtos = await ProdutoService.index();
             return res.status(200).json(produtos);
         } catch (error) {
@@ -15,6 +23,13 @@ export class ProdutoController {
 
     async show(req: Request, res: Response): Promise<Response> {
         try {
+            
+            if(!req.headers.authorization.split(' ')[1]) return res.status(403).json({ error: "Sem autorização!" });
+
+            const validate = jwt.verify(req.headers.authorization.split(' ')[1]);
+
+            if(await validate == null) return res.status(403).json({ error: "Sem autorização!" });
+
             const produto = await ProdutoService.show(Number(req.params.id));
             return res.status(200).json(produto);
         } catch (error) {
@@ -25,6 +40,13 @@ export class ProdutoController {
 
     async create(req: Request, res: Response): Promise<Response> {
         try {
+
+            if(!req.headers.authorization.split(' ')[1]) return res.status(403).json({ error: "Sem autorização!" });
+
+            const validate = jwt.verify(req.headers.authorization.split(' ')[1]);
+
+            if(await validate == null) return res.status(403).json({ error: "Sem autorização!" });
+
             const restaurante = await RestauranteService.show(Number(req.body.id_restaurante));
 
             if (!restaurante) return res.status(404).json({ error: "Restaurante não encontrado" });
@@ -39,6 +61,13 @@ export class ProdutoController {
 
     async update(req: Request, res: Response): Promise<Response> {
         try {
+
+            if(!req.headers.authorization.split(' ')[1]) return res.status(403).json({ error: "Sem autorização!" });
+
+            const validate = jwt.verify(req.headers.authorization.split(' ')[1]);
+
+            if(await validate == null) return res.status(403).json({ error: "Sem autorização!" });
+
             req.body.id_restaurante = undefined;
             const produto = await ProdutoService.update(Number(req.params.id), req.body);
             return res.status(200).json(produto);
@@ -49,6 +78,13 @@ export class ProdutoController {
 
     async delete(req: Request, res: Response) {
         try {
+
+            if(!req.headers.authorization.split(' ')[1]) return res.status(403).json({ error: "Sem autorização!" });
+
+            const validate = jwt.verify(req.headers.authorization.split(' ')[1]);
+
+            if(await validate == null) return res.status(403).json({ error: "Sem autorização!" });
+        
             const produto = await ProdutoService.destroy(Number(req.params.id));
             return res.status(200).json(produto);
         } catch (error) {
